+++ src/HetrickCV.hpp
@@ -62,6 +62,7 @@ extern Model *modelMinMax;
 extern Model *modelPhaseDrivenSequencer;
 extern Model *modelPhaseDrivenSequencer32;
 extern Model *modelPhasorAnalyzer;
+extern Model *modelPhasorBurstGen;
 extern Model *modelPhasorDivMult;
 extern Model *modelPhasorEuclidean;
 extern Model *modelPhasorGates;