+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # HetrickCV Changelog
 
+## 2.3.1 (In Progress)
+- Add Phasor Burst Generator module (in progress).
+- Fix some labels on Phasor Generator.
+- Fix Finish output polyphony on Phasor Generator.
+
 ## 2.3.0
 - Version 2.3.0 introduces a suite of phasor-based sequencing and Phase Distortion synthesis tools.
     - Add Phasor Generator module.